+++ src/main/java/rusty/vanillo/client/ColoredGlints.java
@@ -1,0 +1,5 @@
+package rusty.vanillo.client;
+
+public final class ColoredGlints {
+
+}