+++ src/main/java/rusty/vanillo/enchantment/AbyssalImpulseEnchantment.java
@@ -18,5 +18,4 @@ public class AbyssalImpulseEnchantment extends Enchantment {
     public int getMaxLevel() {
         return 1;
     }
-
 }