+++ src/main/java/rusty/vanillo/enchantment/BanishmentEnchantment.java
@@ -18,5 +18,4 @@ public class BanishmentEnchantment extends Enchantment {
     public int getMaxLevel() {
         return 3;
     }
-
 }